Refactor navigation menu and migrations
- Added "Relatórios" link to the navigation menu
- Changed data types of "horimetro" and "km" columns in the "inspecoes" table from integer to float
- Added PDF button to the "InspecoesTable" component
- Updated label text from "Sim" to "OK" in the "view.blade.php" file
- Added routes for generating PDF reports and viewing reports in the "web.php" file
- Updated input fields for "horimetro" and "km" in the "form.blade.php" file to allow decimal values

+++ database/migrations/2024_01_04_232100_create_inspecoes_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->foreignId('veiculo_id');
             $table->foreignId('modelo_veiculo_id');
             $table->string('uvs')->nullable();
-            $table->integer('horimetro')->nullable();
-            $table->integer('km')->nullable();
+            $table->float('horimetro')->nullable();
+            $table->float('km')->nullable();
             $table->dateTime('data')->nullable();
             $table->string('turno')->nullable();
             $table->string('observacao')->nullable();
